arm: DT: panel-Honami: Remove CABC deferred-on commands
We are currently doing Content-Adaptive Backlight Control
via MDSS features.
It is completely useless for us to enable a worker that
continuously send commands to the DriverIC/panel for its
own CABC: it was also conflicting with QC's CABC.

Why did we have it?
Simply, the original 3.4 copyleft kernel didn't have the QC
adaptive backlight because of various kernel bugs.